# Remove deep copy of indices from shallow()