Close #4311.

This ~still needs a news item~ might need some benchmarking to confirm that the other occurrences of `shallow()` had the same impact as they did on joins. But they should. Those would be
- most setops
- frankv
- dcast
- foverlaps
- duplicated.data.table

And indirectly of course everything that calls these. But the overhead is only really noticeable when there are many calls as in the original report. I don't know when I'll be able to finish this so I'm putting it up as is for now.

In the code there was a reference to `example(merge)` breaking if the sorted attribute wasn't copied. That seems to have been due to a pattern of calling setnames internally in the past. But that pattern hasn't been in use for about five years and nothing currently breaks if I don't make that copy. I do it anyway and have added a test for it as I think it's expected to work, even if `shallow()` isn't exported.