@jangorecki 2020-05-10
This is great, I was recently worrying about using too much of `shallow` and being affected by this issue.