@jangorecki 2020-05-10
We need tests that uses `address` function, those tests need to fail on current master, and pass on this PR.