@tlapak 2020-05-10
Yeah that makes sense. So comparing `address(attr(attr(dt, 'index'), '__a'))` to the shallow copy should do it I think.
Oh yeah, I wasn't too worried, only a bit mystified by nqrecreateindices.c showing up in the report. But I reckon this is due to me working off of the branch before the last merge.