@tlapak 2020-05-22
It's not so much about needing a bigger machine, it's more about finding the time to come up with meaningful benchmarks. Just for `shallow()` it's clear enough so I'll add a small example below. But other operations take so much more time that they drown out the index copying in most situations and the effect only really becomes apparent in loops, similar to the original report. The results aren't massive, but they're there.

```R
shallow <- data.table:::shallow
set.seed(1L)
n <- 1e8
dt1 <- data.table(a=sample(n, n))
setindex(dt1, a)

system.time({
    dt2 <- shallow(dt1)
})
# before
# user  system elapsed 
# 0.11    0.03    0.15 

# now
# user  system elapsed 
#    0       0       0 
```
And for some functions:
```R
library('data.table')

shallow <- data.table:::shallow
set.seed(1L)
n <- 5e6
dt1 <- data.table(a=sample(n, n))
setindex(dt1, a)

dt2 <- data.table(a=sample(1e2, 1e2))
setindex(dt2, a)

gc()
system.time({
  for (i in seq_len(1e2))
    fintersect(dt2, dt1)
})
# before
#   user  system elapsed 
#  72.35    9.45   54.41 

# now
#   user  system elapsed 
#  69.65    8.80   49.03 

system.time({
  for (i in seq_len(1e2))
    duplicated(dt1)
})

# before
#   user  system elapsed 
#  16.25    4.11   14.28 

# now
#   user  system elapsed 
#  17.13    4.53   14.76 

system.time({
  for (i in seq_len(1e2))
    frankv(dt1, 'a')
})
# before
#   user  system elapsed 
#  65.17    7.88   50.44 

# now
#   user  system elapsed 
#  64.72    6.45   48.38 

system.time({
  for (i in seq_len(1e2))
    all.equal(dt1, dt1)
})
# before
#   user  system elapsed 
#  19.09    7.73   28.00 

# now
#   user  system elapsed 
#  15.23    5.46   21.61 
```